refactor: move binding utils to util.ts
